pyproject.toml
[project]
name = "try-astral-typecheck"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.12",
    "pydantic-settings>=2.9.1",
    "sqlalchemy>=2.0.40",
]

[dependency-groups]
dev = [
    "ruff>=0.11.9",
    "ty>=0.0.0a8",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "httpx>=0.27.0",
    "greenlet>=3.2.2",
]
